ovn-tester/ovn_workload: Extract generic methods from Cluster class.
Commit 1380bba13ab34d5fc3f93bf60a30737c95f8d945 moved instantiation
of CMS Cluster class back to the generic ovn_tester.py.

The `create_cluster` function is generic and actually takes almost
identical arguments as the `Cluster` class.  Let's make it part of
the `Cluster` constructor instead!

This change also makes it apparent that a separate CMS plugin class
is not necessary, so let's roll this into the CMS specifc Cluster
class.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>