ci: Add coverage for all address family combinations.
To achieve this, split the CirrusCI job into two tasks:
- one preparing the runtime (for both supported OS versions)
- another one running the tests (defined as a matrix to test all
combinations of IC/non-IC dual/single-stack).

As each task is run in a separate VM instance, we need a way to install
dependencies in the second VM; for that we extend do.sh with a command
option to "install-deps".

Due to the fact that the second set of tasks, low_scale_task, need to
use exactly the runtime built by the commit's runtime_builder_task we
need to explicitly tag all caches with the CirrusCI build ID.  This has
the unfortunate side effect that we can't cache the same runtime across
multiple commits.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>